AsyncNoteLoaderAndParser now notifies main UI if there are no notes to be parsed.
Transformed AsyncNoteLoaderAndParser from a runnable to a normal class (listing files is not time consuming enough for the added complexity). It now sends an empty message to ui.Tomdroid if no notes are in the tomdroid/ folder.
Added NO_NOTES msg type (int constant) to Note
NoteCollection uses modified AsyncNoteLoaderAndParser